backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
langchain>=0.3,<0.4
langchain-google-genai>=2.0
langchain-community>=0.3
langchain-text-splitters>=0.3
faiss-cpu>=1.9
pypdf>=5.0
python-dotenv>=1.0
python-multipart>=0.0.18
pydantic>=2.0
